refactor(modules/history): move complex logic to service
