+++ src/modules/history/history.repository.spec.ts
@@ -1,18 +1,14 @@
 import { Test } from '@nestjs/testing'
 import { DataSource } from 'typeorm'
 import { MockProxy, mock } from 'vitest-mock-extended'
-import { PlayHistory } from '@spotify/web-api-ts-sdk'
 
 import { HistoryRepository } from './history.repository'
 import { History } from './history.entity'
-import { HistoryTracksRepository } from './tracks'
-import { HistoryTrack } from './tracks'
 
 describe('HistoryRepository', () => {
   const userId = 'userId'
 
   let historyRepository: HistoryRepository
-  let historyTracksRepository: HistoryTracksRepository
 
   let historyEntityMock: MockProxy<History>
 
@@ -25,19 +21,12 @@ describe('HistoryRepository', () => {
             createEntityManager: vi.fn(),
           },
         },
-        {
-          provide: HistoryTracksRepository,
-          useValue: {
-            createHistoryTracksFromPlayHistory: vi.fn(),
-            findLastHistoryTrackByUser: vi.fn(),
-          },
-        },
+
         HistoryRepository,
       ],
     }).compile()
 
     historyRepository = module.get(HistoryRepository)
-    historyTracksRepository = module.get(HistoryTracksRepository)
 
     historyEntityMock = mock<History>()
   })
@@ -61,117 +50,6 @@ describe('HistoryRepository', () => {
           id: userId,
         },
       },
-    })
-  })
-
-  describe('updateOrCreateHistoryByUser', () => {
-    let historyTrack: MockProxy<HistoryTrack>
-    let historyTracks: MockProxy<HistoryTrack[]>
-    let playHistoryMock: MockProxy<PlayHistory[]>
-
-    beforeEach(() => {
-      historyTrack = mock<HistoryTrack>()
-      historyTracks = [historyTrack]
-      playHistoryMock = mock<PlayHistory[]>()
-
-      historyEntityMock.historyTracks = historyTracks
-    })
-
-    test('should create new history', async () => {
-      const findHistoryByUserSpy = vi
-        .spyOn(historyRepository, 'findHistoryByUser')
-        .mockResolvedValue(null)
-      const createHistoryTracksFromPlayHistorySpy = vi
-        .spyOn(historyTracksRepository, 'createHistoryTracksFromPlayHistory')
-        .mockResolvedValue(historyTracks)
-      const createSpy = vi
-        .spyOn(historyRepository, 'create')
-        .mockReturnValue(historyEntityMock)
-      const saveSpy = vi
-        .spyOn(historyRepository, 'save')
-        .mockResolvedValue(historyEntityMock)
-
-      expect(
-        await historyRepository.updateOrCreateHistoryByUser(
-          historyEntityMock.user,
-          playHistoryMock
-        )
-      ).toEqual(historyEntityMock)
-
-      expect(findHistoryByUserSpy).toHaveBeenCalledWith(
-        historyEntityMock.user.id
-      )
-      expect(createHistoryTracksFromPlayHistorySpy).toHaveBeenCalledWith(
-        playHistoryMock,
-        historyEntityMock.user
-      )
-      expect(createSpy).toHaveBeenCalledWith({
-        user: historyEntityMock.user,
-        historyTracks,
-      })
-      expect(saveSpy).toHaveBeenCalledWith(historyEntityMock)
-    })
-
-    test('should update existing history if historyTracks is not empty', async () => {
-      const findHistoryByUserSpy = vi
-        .spyOn(historyRepository, 'findHistoryByUser')
-        .mockResolvedValue(historyEntityMock)
-      const findLastHistoryTrackByUserSpy = vi
-        .spyOn(historyTracksRepository, 'findLastHistoryTrackByUser')
-        .mockResolvedValue(historyTrack)
-      const createHistoryTracksFromPlayHistorySpy = vi
-        .spyOn(historyTracksRepository, 'createHistoryTracksFromPlayHistory')
-        .mockResolvedValue(historyTracks)
-      const saveSpy = vi
-        .spyOn(historyRepository, 'save')
-        .mockResolvedValue(historyEntityMock)
-
-      expect(
-        await historyRepository.updateOrCreateHistoryByUser(
-          historyEntityMock.user,
-          playHistoryMock
-        )
-      ).toEqual(historyEntityMock)
-
-      expect(findHistoryByUserSpy).toHaveBeenCalledWith(
-        historyEntityMock.user.id
-      )
-      expect(findLastHistoryTrackByUserSpy).toHaveBeenCalledWith(
-        historyEntityMock.user.id
-      )
-      expect(createHistoryTracksFromPlayHistorySpy).toHaveBeenCalled()
-      expect(saveSpy).toHaveBeenCalledWith(historyEntityMock)
-    })
-
-    test('should update existing history if historyTracks is empty', async () => {
-      const findHistoryByUserSpy = vi
-        .spyOn(historyRepository, 'findHistoryByUser')
-        .mockResolvedValue(historyEntityMock)
-      const findLastHistoryTrackByUserSpy = vi
-        .spyOn(historyTracksRepository, 'findLastHistoryTrackByUser')
-        .mockResolvedValue(null)
-      const createHistoryTracksFromPlayHistorySpy = vi
-        .spyOn(historyTracksRepository, 'createHistoryTracksFromPlayHistory')
-        .mockResolvedValue(historyTracks)
-      const saveSpy = vi
-        .spyOn(historyRepository, 'save')
-        .mockResolvedValue(historyEntityMock)
-
-      expect(
-        await historyRepository.updateOrCreateHistoryByUser(
-          historyEntityMock.user,
-          playHistoryMock
-        )
-      ).toEqual(historyEntityMock)
-
-      expect(findHistoryByUserSpy).toHaveBeenCalledWith(
-        historyEntityMock.user.id
-      )
-      expect(findLastHistoryTrackByUserSpy).toHaveBeenCalledWith(
-        historyEntityMock.user.id
-      )
-      expect(createHistoryTracksFromPlayHistorySpy).toHaveBeenCalled()
-      expect(saveSpy).toHaveBeenCalledWith(historyEntityMock)
     })
   })
 })
