+++ src/modules/history/index.ts
@@ -2,3 +2,4 @@ export * from './history.entity'
 export * from './history.repository'
 export * from './history.module'
 export * from './history.scheduler'
+export * from './history.service'
